Protocol test hspec
